refactor(platform): move the in-memory automation store to test support
lib/engine/store/memory.ts shipped in the production tree, type-checked
and bundled as a second store implementation next to the real pg store,
yet its only importers were five test files — no host, script, or
public export ever used it.

Move it to lib/engine/selftest/memory-store.ts beside the selftest that
depends on it, repoint the test imports, and drop `store/` from the
purity test's pure-layer list (the directory no longer exists; the
test-support module is host code like the tests themselves).

Finding: lib-engine-6.
